requirements.txt
numpy==1.21.6
pandas==1.3.5
psycopg2==2.9.6
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
